+++ packages/pwa-kit-dev/src/configs/webpack/config.js
@@ -508,6 +508,9 @@ const ssr = (() => {
             .extend((config) => {
                 return {
                     ...config,
+                    ...(process.env.PWA_KIT_SSR_SOURCE_MAP === 'true'
+                        ? {devtool: 'source-map'}
+                        : {}),
                     // Must *not* be named "server". See - https://www.npmjs.com/package/webpack-hot-server-middleware#usage
                     name: SSR,
                     entry: `.${EXT_OVERRIDES_DIR}/app/ssr.js`,